pdf_related_requirements.txt
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.10
brave-search==0.1.8
Brotli==1.1.0
camelot-py==0.11.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
cryptography==42.0.2
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
dataclasses==0.8
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
de-dep-news-trf==3.7.2
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
emoji==2.10.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
filetype==1.2.0
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.48.1
fsspec==2024.2.0
gevent==24.2.1
ghostscript==0.7
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
JPype1==1.5.0
jsonpath-python==1.0.6
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
layoutparser==0.3.4
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
munkres==1.1.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest_asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0
openpyxl==3.1.2
packaging==23.2
pandas==1.5.3
parso==0.8.3
pathy==0.10.2
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pdftopng==0.2.3
pexpect==4.9.0
pickleshare==0.7.5
pikepdf==8.12.0
pillow==10.2.0
pillow_heif==0.15.0
pip==23.3.1
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.42
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.1
pypdf==3.17.4
PyPDF2==2.10.5
pypdfium2==4.27.0
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.1.1
pytest-asyncio==0.23.5.post1
python-dateutil==2.8.2
python-docx==1.1.0
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
scipy==1.12.0
sentencepiece==0.2.0
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
smart-open==5.2.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.2
spacy-alignments==0.9.1
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.4
srsly==2.4.8
stack-data==0.6.2
sympy==1.12
tabula-py==2.9.0
tabulate==0.9.0
tenacity==8.2.3
texttable==1.7.0
thinc==8.2.2
timm==0.9.12
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.14.1
typer==0.9.4
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2024.1
unicodedata2==15.1.0
unstructured==0.12.4
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==2.2.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
websearch-python==1.2.1
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.16.0
XlsxWriter==3.1.9
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
